test: adds response test cases to "validateElement" tests
